feat: enhance Chat component and job monitoring hooks for improved background job handling - Update Chat component to prevent duplicate job processing and clear states after successful processing - Modify useTriggerJobMonitor to reset states for new jobs and ensure only current job updates are processed - Adjust useTriggerJobMonitorWithAuth to always pass triggerJobId and callbacks for consistency - Introduce BackgroundJobStatus interface for better job status management - Remove obsolete backgroundJobService interface to streamline codebase
