--- frontend/internal-packages/agent/src/chat/workflow/services/backgroundJobService.ts
@@ -1,8 +1,0 @@
-export interface BackgroundJobStatus {
-  jobId: string
-  status: 'pending' | 'processing' | 'completed' | 'failed'
-  generatedAnswer?: string
-  error?: string
-  createdAt: string
-  completedAt?: string
-}
